BUG#22974099: MEMORY LEAK ON RELAY_LOG_INFO::UNTIL_OPTION
Class Relay_log_info contains a Until_option instance, which is
allocated at start_slave(), that may not be deleted in some
conditions, like when the until condition is not reached.
This may cause a memory leak on Relay_log_info::until_option.

To solve the above mentioned leak, we always delete Until_option
instance on Relay_log_info destructor.