Fixed bug#21982792: TOO MUCH SPAM: 'GOT ERROR 1 WHEN READING TABLE'
When reading one row by reference scan, there are two steps: 1) copy the
referenced value to the indexed field 2) use the indexed field value to scan
the index. During step 1, we might meet warnings such as a type conversion
warnings. In strict mode, these warnings will be treated as errors. However in
step 1, it misses to check the thd->is_error to see whether there is such kind
of error. The missing error handling in the server results in the server goes
forward like no error. At the same time, it will results in some abnormal
behavior next, such as update_generated_read_fields will throw out an error
because it checks thd->is_error(). This is unreasonable.

This patch fixes such kind of problems. The thd->is_error() will be detected
before calling update_generated_read/write fields.  The places where
the->is_error() is checked are listed as the below:
1. fill_record() checks after save_in_field()
2. cp_buffer_from_ref() does after copy()
3. copy_data_between_tables() does after invoke_do_copy()
4. do_updates() does after invoke_do_copy()
5. JOIN_CACHE::generate_full_extensions checks after check_match()

This patch causes changes to the test results from i_main.sp. This is caused
by the new error handling which makes the server return earlier when an error
has occurred.