WL#7069: Provide data dictionary information in serialized form
This worklog provides a the data dictionary information for schemas,
tables and tablespaces in human-readable form (JSON). For storage
engines which don't support storing the serialized dictionary
information (SDI) in tablespaces (all SEs except Innodb), the SDI is
written to a file in datadir.

For Innodb the the SDI will be stored in a tablespace, but this
remains disabled for now, as it depends on worklog number 7141.

The SDIs can be manually edited and will be used for import and
disaster recovery (to be provided worklog number 7524 and 7412).

Note that the name of the SDI files includes the Object_id from the
data dictionary, which is a value created by an auto-increment column
in the corresponding data dictionary table. This in turn means that
the name of an SDI file created by an MTR test will not necessarily be
the same each time the test is run as it depends on how many dd
operations have been performed on the server prior to the test being
run. In particular it will not be same when running a test by itself as
when running the test as part of a suite. This does not cause any
problems, except in cases where the test includes the content of
datadir in its output. Then a --replace_regexp must be used to make
the output of the test stable. E.g.:

--replace_regex /_[0-9]+\.SDI/_XXX.SDI/
--list_files $MYSQLD_DATADIR/mysqltest/