Bug#22900110: SIMPLIFY RAPID_JSON_HANDLER
Rapid_json_handler has its own internal data structures to represent
the partially built DOM for the JSON document that is being parsed.
These data structures need to be converted to a Json_dom graph before
the result can be returned to caller.

This patch changes Rapid_json_handler so that it instead builds the
Json_dom graph directly. This lets us get rid of most of the internal
data structures that keep track of the parser's state, and reduces the
amount of work required to build the resulting Json_dom graph.

The patch also refactors the constructor of Json_string to avoid
unnecessary copying of std::string objects in some cases.