Bug#22973279 IMPROVE ERROR HANDLING FROM CALL TO STORE_KEY::COPY() IN CP_BUFFER_FROM_REF
The function store_key::copy() returns an enum value, but in
cp_buffer_from_ref() this return value is treated as if it is a bit
vector. If the first bit is set, then this is treated as
store_key::copy() returns an error.

This patch changes cp_buffer_from_ref() to check that
store_key::copy() returns the enum value store_key::STORE_KEY_OK
instead treating the return value as a bit vector.