Bug#23028144 CORE STACK NOT SHOWING ASSERTION
This is a regression from
Bug#22996442 INNODB: MAKE UNIV_DEBUG DEPEND ON DBUG_OFF, AND REPLACE
REFERENCES TO DBUG_OFF.

When ut_ad() was made an alias of DBUG_ASSERT(), which in turn is an
alias of assert(), the assertion expression would not necessarily
any longer be in the call stack, depending on the implementation of
assert().

The fix is to make ut_ad() an alias of ut_a() like it used to be,
but also improve ut_dbg_assertion_failed() so that it displays
the file name, line number and assertion expression in a single line,
for easier viewing of the error logs.

ut_dbg_assertion_failed(): Use sql_print_error() to display the
file name, line number and message in a single line, and the thread ID
in a subsequent line.

Adjust some tests to suppress warnings about the sql_print_error()
messages for assertion failures, including ib::fatal().

RB:12295
Reviewed-by: Jimmy Yang <jimmy.yang@oracle.com>