Bug#23036141 - IBD2SDI DOESN'T WORK IF ANY DATAFILE IS SMALLER THAN PREVIOUS DATAFILE
Problem:
--------
When multi-file tablespace is passed to ibd2sdi, it fails to read the SDI
copies.

For example: if ibdata1 is 50M and ibdata2 is 5M, the following will fail
ibd2sdi --read=0 ibdata*

ibd2sdi calculates the size of tablespace wrongly. When new add
file is added, it takes second size and overwrites the existing size.

Fix:
----
Do not track last page num of tablespace, instead account the total number of pages
correctly.

Reviewed-By : Bin Su <bin.x.su@oracle.com>
RB : 12577