Bug #22447330: INTRODUCE NOT_VALGRIND_SERVER.INC TO ENABLE VALGRIND FOR CLIENTS
An include file 'not_vaglrind_server.inc' has been introduced through which
tests can specify that they are not to be run when valgrind is enabled for
the server. This allows such tests to run with the '--valgrind-clients' MTR
option. An environment variable $VALGRIND_SERVER_TEST is being set to
indicate whether valgrind is enabled for mysqld.

Reviewed by :
Sayantan Dutta <sayantan.dutta@oracle.com>
RB : 12257