Bug#22884631 FAILING ASSERTION: TRX->DUPLICATES == 0
Issue:
======
When server uses a TABLE* created by one thread in another,
we generally update the trx associated in
ha_innobase::external_lock() or the first handler call (we
have update_thd() in many handler functions).

With invalid statements like wrong number of fields for an
insert,it is possible server opens a table using the used
TABLE* handle, then detects the wrong fields and skips the
external_lock() call or other handler calls like extra() or
info().

ha_innobase::end_stmt() is directly called to end the
statement. At this time, the trx is still not updated and
this causes the assert failure.

Fix:
====
In ha_innobase::end_stmt(), we first check if the
prebuilt->trx has been updated or not. We skip the assert
if it has not been updated.

RB: 12277
Reviewed-by: Satya Bodapati <satya.bodapati@oracle.com>