Bug#22819614 OUT OF RANGE SRIDS ARE SILENTLY CAST TO UINT32
Problem: Geometry import functions that take an SRID parameter cast it
to uint32 without warning or error. Negative values or values larger
than UINT_MAX32 are silently converted to a number within range.

Fix: In all functions that take the SRID as a parameter, check that
it's within the range of a uint32 and return ER_DATA_OUT_OF_RANGE if
it isn't.

This is also applied to the GeoJSON and geohash functions that already
checked that the parameter was within range, but returned a different
error code.