Bug #22893130: BOOTSTRAP FAILS IF INNODB USES RAW DEVICE
When using a raw device, innodb will execute ALTER statements
using the COPY algortihm, not INPLACE. At the end of
bootstrapping the DD, we add a foreign key to dd.character_sets
using ALTER TABLE. This FK is cyclic in the sense that the table
referred has another FK referring back to dd.character_sets,
and thus, the FK cannot be added as part of the CREATE TABLE
statement. Normally, this ALTER statement is executed using the
INPLACE algorithm, but when running with a raw device, innodb
decides to do it using the COPY algorithm.

The objects representing the DD tables are sticky in the cache
to prevent them from being evicted (because once evicted, no
cache misses can be handled, since the DD objects must be present
in order to handle the cache miss). At the end of bootstrapping
the DD, we make sure all DD objects are indeed sticky. A bug in
ALTER COPY failed to retain stickiness, and this caused a failing
assert that made bootstrap fail when running with a raw device.

When fixing this bug at the SQL layer, we still have the issue
that ALTER INPLACE is rejected by innodb when running with a raw
device. This will be a problem at a later stage when ALTER COPY
is rejected on the hardcoded DD tables.

This patch is an intermediate solution, and will be replaced
by a more permanent fix in WL#7743.

Unrelated to the above, the patch also removes an unused
constant.