Bug#22720753: ENABLE OR REMOVE THE DISABLED TESTS UNDER ENGINES SUITE
There are several tests in the engines suite which are in disabled
state from a long time. They have been evaluated whether they are
required, or are removed. The tests which have been retained have
been fixed to conform to the current behaviour.

Reviewed by:
Amit Bhattacharya <amit.bhattacharya@oracle.com>
RB #11957