Bug#22783011: assertion failed: base_select != removed_select && aggr_select
The problem query contains an aggregation clause in the WHERE clause of
a derived table definition. However, it is invalid to aggregate an
expression in the WHERE clause, so we have to assume that the aggregation
takes place in an outer query block. But derived tables do not allow
outer references, so aggregation cannot take place there either, thus it
is reasonable to reject this query as invalid.

Fixed by clearing LEX::allow_sum_func before preparing the query
expression of a derived table. This will ensure that any aggregation
happens within the derived table only.