BUG#17944190 : MYSQLTEST CANNOT PARSE PERL BLOCK INSIDE IF THAT EVALUATES TO FALSE
Description :
=============
mysqltest fails to parse a 'perl' statement inside an 'if'
statement, if the 'if' condition evaluates to false. It can
handle the case when the 'if' condition evaluates to true.

Issue :
=======
do_perl() is the method which parses the perl block. When
'if' condition evaluates to false, do_perl() method is not
called and mysqltest conideres perl statements as SQL
statements and parses till it finds the delimiter i.e ';'.

Fix :
=====
Call do_perl() to parse the perl block but doesn't execute
the perl block if 'if' condition evaluates to false.

Reviewed-by: Sayantan Dutta <sayantan.dutta@oracle.com>
RB: 12499