WL#6378: New data dictionary
Follow-up patch: Refactor Collection and Iterator
usage in the new data dictionary code base.

This patch removes the dd::Iterator interface. Dictionary
collections instead have iterators that conform to std::iterator.
This means that they can be used for STL algorithms and
range-based for-loops. Instead of returning an iterator,
type classes now return a const reference to dd::Collection
which implements begin() and end().

The patch combines dd::Base_collection and dd::Collection
into one class. The internal std::vector now contains pointers
to the actual Impl classes, rather than to dd::Collection_item.
This means that dd::Collection_item could be removed. Instead,
a typedef in the abstract type classes point to the impl class.

dd::Collection_item_factory is also removed. Instead, it
is expected that impl clases implement static restore_item()
and clone() functions.

Set and Enum elements are now stored in a single collection
for columns and parameters. This is possible since columns
and parameters cannot have both types at the same time.

The new iterators do not support skipping of hidden items.
Instead this is expected to be handled by the API user by
checking is_hidden() for the individual items as needed.

Partitions are no longer sorted by default. This must again
now be handled by the API caller. This makes it possible
to handle partitions similar to other types. The dd_partition
unit test was removed as it only tested sorting.

The patch also removes a lot of const_casts and dynamic_casts.