Bug #22276843: INCORRECT BEHAVIOR OF SET VARIABLE WITH                SUBQUERY AND ROW CONSTRUCT
ISSUE:
------
The root cause here is that SET statement locks the tables
at the beginning. This allows the optimizer to read tables
and evaluate subqueries during the PREPARE phase.

The current query is of the form:
SET @x:= (SQ1 ORDER BY (SQ2 IN (SQ3)));

- The IN-to-exists (or subquery materialization, depending
on the hint) transform is applied to the inner subquery.

- Since this is a SET statement, SQ1 is also considered as
a subquery. ORDER BY inside a subquery is redundant and is
removed by remove_redundant_subquery_clauses(). This does
clean_up_after_removal() on SQ3, which unplugs SQ2 and its
children (SQ3) from the chain of units and calls
invalidate() on them.

- The crash is because, unlike cleanup() of item_subselect
and item_singlerow_subselect, item_in_subselect::cleanup()
tries to access first_select(). It does that because
exec_strategy is EXEC_MATERIALIZATION or EXEC_EXISTS. In
this case, SQ3 is evaluated and then it is invalidated for
the reasons mentioned above. Normally, for an invalidated
unit ("unplugged" unit) the containing Item_in_subselect
hasn't been evaluated yet so exec_strategy is
EXEC_EXISTS_OR_MAT and things are fine.

SOLUTION:
---------
Currently optimizer the steps for a SET statement are:
1) check_table_access
2) open_and_lock_tables
3) sql_set_var
   a) check
      i) fix_fields
      ii) ......


Notice here that the locking of the tables precedes the
fix_fields. As mentioned above this causes the optimizer
to assume that, with locked tables, the output of the
subquery will be constant.

To remedy this issue, we should make sure that the SET
statements don't lock the tables before fix_fields. In this
fix, the locking of tables is delayed until after
fix_fields, hence avoiding the above mentioned problems.
The new steps will be:

1) check_table_access
2) open_tables_for_query
3) sql_set_var
   a) resolve
      i) fix_fields
      ii) ......
   b) lock_tables
   c) check
   .......

Also, I'm moving the DBUG_EXECUTE_IF
"open_tables_for_query__out_of_memory" to
Sql_cmd_insert::mysql_test_insert and renaming it. Since
open_tables_for_query will be used for SET statements,
this test can be moved to an earlier insert-related function.