BUG#20970917 : MTR --MARK-PROGRESS WRITES STDOUT IF RESULT FILE MISSING
Description:
============
When the --mark-progress is used with MTR and the test case does not
have a result file, the progress is written to stdout, mixed with the
result log. This makes it unusable.

MTR --mark-progress is very useful when developing/debugging test cases,
since if the test hangs you can check where it has hung. When developing
a test case it is often practical to do it without recording a result file.

Issue:
======
MTR uses result file name to create the progress file. If result file
doesn't exist, it prints the progress report on console along with the
test output.

Fix :
====
When result file doesn't exist, use cur_file->file_name variable to create
the progress file.

Reviewed-by: Sayantan Dutta <sayantan.dutta@oracle.com>
Reviewed-by: Srikanth B R <srikanth.b.r@oracle.com>
RB: 12598