wl#7587:Migrate the testcases InnoDB as a default storage engine
problem :

Some tests are in main suite, fails when runing with InnoDB as a default storage engine,
So currently, these tests are being run with have_myisam.inc

Solution:

Run the tests by Removing inc file : have_myisam.inc and analyze each one of the failures.
Record the result files cautiously, file a bug in case of failures, which are un-expected
as per the storage engine wise.

RB : 12028