Bug#23477340 I_MAIN.INVALID_PACKET SKIPS ON DEBIAN PLATFORM
Issue: MTR test i_main.invalid_packet skips on debian platform even if netcat
utility is installed on the server. This is because of an identified bug in
"have_util_nc.inc" file. The inc file checks for the string "usage" in the
output of the command nc -h or netcat -h

However, the output on Debian does not matches the string "usage".

Fix: The logic is correct which works on all platform.

Reviewed by: Pavan Naik <pavan.naik@oracle.com>
RB: 12823