BUG#23474669 : MTR: PRESENCE OF CHECK-WARNINGS.LOG IN VARDIR/TMP CAUSES CHTEST FAILURE
Description :
=============
Presence of check-warnings.log in VARDIR/tmp causes chtest to
fail due to a result content mismatch in the check for the
directory's contents. This seems to be a regression introduced
by the fix for Bug#18274766.

The CHECK TESTCASES module of chtest checks for the contents of
$MYSQLTEST_VARDIR/tmp directory and finds an extra file present
leading to a mismatch of files.

Fix :
=====
Since check-warnings.test doesn't produce any output, delete the
check-warnings.log file generated after the check-warnings test
run is completed.

Reviewed-by: Srikanth B R <srikanth.b.r@oracle.com>
RB: 12822