Bug #23212765: ASSERTION FAILED: SP_RESULT_FIELD
ISSUE:
------
This problem occurs under the following conditions:
1) A function returns a json value.
2) This function is called as part of a case statement in a
   procedure.

SOLUTION:
---------
Once the case_value is evaluated, it is stored in a
Item_cache_* object. The function values are then cleaned
up. The Item_cache_* object in turn becomes part of a
Item_func_eq object (to test the equality with the value in
the WHEN clause).

item.h:
While preparing the WHEN clause,
Item_cache_json::result_type is called. This function tries
to access the example object, which is the original object
from which it was created. But the attributes of the
example object have already been cleaned up after the
function was evaluated.

There is really no need for Item_cache_json::result_type to
lookup the example object. The result type for this object
will always be STRING_RESULT.


item_func.cc:
1) Item_func_sp::execute is supposed to return true only
   when there is an error. But it returns null_value flag,
   which is just wrong. Also, this leads to some incorrect
   treatment of the return value from execute() function
   in some places.

2) One such place is Item_func_sp::val_json. A null value
   is treated as equivalent to an error and true is
   returned here too. This has been corrected.

These changes influence whether the value_cached flag is
set in Item_cache_json::cache_value. Here inadvertently,
this is set to false, despite the evaluation, we will see
incorrect behavior at a later stage.

item_func.h:
For Item_func_sp::val_str, a null value should be returned
whether there was an error from the execute or the result
of the function was actually a null value. Similar changes
are made for other val_* and get_time/date functions.