BUG#22841097 : TESTS FAIL WITH "MERGE MISMATCH WARNINGS IN GCOV"
Issue :
=======
Tests occasionally fail on GCOV platform in PB2 with merge mismatch
warning.

Fix :
=====
For GCOV runs when there is a result difference, run the output
through a "sed -i '/gcda:Merge mismatch for function/d'" and try again.

Filtering is done in check_result() before calling compare_files(),
so the latter doesn't have to guess whether or not to do the filtering.

Reviewed-by: Bjorn Munch <bjorn.munch@oracle.com>
RB: 12703