Bug#23540008 SAFE GUARD FOR CHARSET_INFO RETURNED FROM GET_CHARSET
When executing a SELECT from tables:
- performance_schema.events_statements_current
- performance_schema.events_statements_history
- performance_schema.events_statements_history_long

the code reads data that can be concurrently written to.

This race condition is expected (performance schema data buffers are lock
less), but the code is not robust enought.

In particular, the character set for the sql query text may be invalid.

Before this fix, this condition could cause a crash.

With this fix, reading an invalid character set will truncate the SQL
TEXT column.