WL#8980 - Move UDF Meta-data table from MyISAM to the           Transactional Storage Engine InnoDB.
The UDF metatable mysql.func from MyISAM is migrated to the
InnoDB storage engine by this patch. Errors from storage engine
and OOMs are handled by full rollback. The code acquires strong
SNRW and SRO locks on mysql.func table (i.e it table-level lock of
MyISAM) to keep the same operational semantics as when the system
table was using MyISAM. THR_LOCK_udf scope is made minimal and
it protects the in-memory hash data struture.