Bug #23748219: MAIN.FILE_CONTENTS TEST BREAKS IF INSTALLED INTO DIRECTORY CONTAINING /LIB
Don't remove /lib from $MYSQL_BINDIR when searching for INFO_{SRC,BIN}.

This was required when we started from $MYSQL_LIBDIR, but in
the fix for #11751935, this was changed to $MYSQL_BINDIR.
So now, this is not only unneccessary, but breaks the main.file_contents
test when run from a directory with /lib in it (e.g. /var/lib/jenkins).

Change-Id: I59b7e4fa77cb102852b834dd5ad0fb1255ec6e5a