Bug #23509085 FAILURE IN I_INNODB.INNODB_BUG16097753 ON WINDOWS
Test i_innodb.innodb_bug16097753 fails on Windows 10 and 7 as a
consequence of the fix to BUG#22305994 TIMESTAMP DATA TYPE PRECISION
IS NOT ACCURATE UPTO NANO SECONDS ON WINDOWS. The failure is caused
by the difference in resolution between the timing methods used
within the innobase and sql libraries that is introduced by the fix
for BUG#22305994.

The Windows time measurement API used for microsecond resolution by
both the innobase and sql libraries prior to the fix for BUG#22305994
was the GetSystemTimeAsFileTime API.  After the fix for BUG#22305994,
the sql library uses the QueryPerformanceCounter API which has a
higher resolution.

This patch fixes the problem by introducing a new function
ut_high_res_usectime into the innobase library that (on Windows) uses
the QueryPerformanceCounter API in a similar fashion to its use in
my_micro_time in the sql library.

On non-Windows platforms, ut_high_res_usectime is a synonym for
ut_usectime.

Reviewed by: Sunny Bains (sunny.bains@oracle.com)
RB: 13063