Bug 1582456: Incremental restore in XB 2.4 requires 1+GB of RAM
With upstream change of `UNIV_PAGE_SIZE_MAX' from 16K to 64K, the size
of `incremental_buffer' in `xtrabackup_apply_delta' becomes 1G. In fact,
we don't need to allocate such a large buffer for smaller pages. What we
really need is `page_size / 4' pages.

The fix doesn't address the issue for tablespaces with 64K page size,
but it works better in default case.