WL#7730: MTR internal check failures should be detected and reported on PushBuild
Summary:
         MTR runs a validation check after each test is run, the validation process checks
for any residue left by the server after the test is completed.

In order to validate this MTR takes a copy of the state of the server before
starting the test and a copy of the state of the server after the test.
These two states are then matched against each other to check if any residue was
left by the test.
A typical example would be to see if the test had created any database in the
server while running but did not remove it.

On such occasions , MTR would find out that the test had although passed but left
a residue.

MTR would although pass the test but would complain(warning message) , that the
test had some leftovers.

Resolution:
   A new option --failcheck-testcases has been provided which defaults to OFF.
If switched ON , it would fail the test cases that fail in checktestcase.

Code Review By:
    Bjorn Munch <bjorn.munch@oracle.com>
     Pavan Naik <pavan.naik@oracle.com>